StudKempYandex2025/Project/tg-bot/requirements.txt
python-telegram-bot==20.5
httpx>=0.23.0,<1.0